pyproject.toml
[project]
name = "mid-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.2",
    "fastapi>=0.135.1",
    "faster-whisper>=1.2.1",
    "httpx>=0.28.1",
    "langchain>=0.3.15,<0.4.0",
    "langchain-community>=0.3.0,<0.4.0",
    "langchain-core>=0.3.31,<0.4.0",
    "langchain-openai>=0.3.0,<0.4.0",
    "langchain-text-splitters>=0.3.0,<0.4.0",
    "langgraph>=0.2.60",
    "moviepy==2.2.1",
    "openai>=1.50.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
    "tavily-python>=0.7.22",
    "uvicorn>=0.41.0",
    "whisper>=1.1.10",
    "youtube-transcript-api<1.0.0",
    "yt-dlp>=2026.3.3",
]
